requirements.txt
# Minimal requirements for Vercel deployment
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pandas==2.1.3
plotly==5.17.0
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0